Refigured members, and fixed errors from previous pull request
