+++ app/data/Members/members.ts
@@ -1,0 +1,34 @@
+const teamMembers = [
+  {
+    name: "Jimmy Munoz",
+    role: "Project Manager",
+    image: "/team/Jimmy.jpg",
+  },
+  {
+    name: "Morad Irshaid",
+    role: "Backend Developer",
+    image: "/team/Morad.jpg",
+  },
+  {
+    name: "Cade Jordan",
+    role: "Frontend Developer",
+    image: "/team/Cade.png",
+  },
+  {
+    name: "Russell Kondoker ",
+    role: "Backend Developer",
+    image: "/team/Russell.jpg",
+  },
+  {
+    name: "Ihsan Sarwar",
+    role: "Frontend Developer",
+    image: "/team/Ihsan.jpg",
+  },
+  {
+    name: "Tegh Gill",
+    role: "Backend Developer",
+    image: "/team/Tegh.jpg",
+  },
+];
+
+export default teamMembers;
